Move obstacle into its own module
